requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
python-multipart
pymupdf